spark/requirements-spark.txt
boto3==1.35.14
botocore==1.35.14
coverage==7.6.0
cx_Oracle==8.3.0
et_xmlfile==2.0.0
googleapis-common-protos==1.56.4
grpcio==1.56.0
grpcio-status==1.56.0
ldap3==2.9.1
openpyxl==3.1.5
pandas==2.3.0
protobuf==4.25.8
py4j==0.10.9.7
pyarrow==16.1.0
PyYAML==6.0.3
pyspark==3.5.4
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.35
tqdm==4.67.1
tzdata==2025.2
minio==7.2.18
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.10.5
cffi==2.0.0
et_xmlfile==2.0.0
numpy==2.3.1
openpyxl==3.1.5
pycparser==2.23
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
google-api-python-client>=2.130.0
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
pdfplumber>=0.11.0